+++ src/main/java/org/davidfabio/game/BulletPlayer.java
@@ -3,12 +3,12 @@ package org.davidfabio.game;
 import com.badlogic.gdx.graphics.Color;
 
 public class BulletPlayer extends Bullet implements Attacker {
-    private float attackPower = 1.0f;
+    private int attackPower = 1;
     private float width = 8;
     private float height = 32;
 
-    public float getAttackPower() { return this.attackPower; }
-    public void setAttackPower(float newAttackPower) { this.attackPower = newAttackPower; }
+    public int getAttackPower() { return this.attackPower; }
+    public void setAttackPower(int newAttackPower) { this.attackPower = newAttackPower; }
 
     @Override
     public void init(float x, float y, float scale, float moveSpeed, float angle, Color color) {
