+++ src/main/java/org/davidfabio/game/EnemyChaser.java
@@ -10,7 +10,7 @@ public class EnemyChaser extends Enemy {
 
 
     @Override
-    public void init(float x, float y, float scale, float moveSpeed, float newInitialHealth, Color color) {
+    public void init(float x, float y, float scale, float moveSpeed, int newInitialHealth, Color color) {
         super.init(x, y, scale, moveSpeed, newInitialHealth, color);
 
         float[] vertices = new float[] {
